correction

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -3,20 +3,26 @@
    with a name containing the letter a
    from the database hbtn_0e_6_usa
 """
-from model_state import Base, State
+
+from sys import argv
+from model_state import State, Base
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
-from sys import argv
 
 if __name__ == "__main__":
+
     db_url = "mysql+mysqldb://{}:{}@localhost:3306/{}".format(
-            argv[1], argv[2], argv[3])
+        argv[1], argv[2], argv[3])
+
     engine = create_engine(db_url)
     Session = sessionmaker(bind=engine)
+
     session = Session()
-    states = session.query(State).filter(State.name.like("%a%")).all()
-    for state in states:
-        session.delete(state)
+
+    states = session.query(State).filter(State.name.contains('a'))
+    if states is not None:
+        for state in states:
+            session.delete(state)
 
     session.commit()
     session.close()
